requirements.txt
aliyun-python-sdk-core==2.13.3
aliyun-python-sdk-kms==2.16.0
amqp==5.1.1
annotated-types==0.6.0
anyio==3.7.1
arrow==1.2.3
asgiref==3.6.0
autopep8==2.0.2
cairocffi==1.5.1
CairoSVG==2.7.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
crcmod==1.7
cryptography==40.0.2
cssselect2==0.7.0
defusedxml==0.7.1
distro==1.8.0
Django==4.1.7
django-redis==5.2.0
django-simpleui==2023.8.28
docopt==0.6.2
exceptiongroup==1.2.0
fastapi==0.105.0
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.4
ijson @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_5c1d7215-d7b6-4d2c-8bf6-886f14b2a0259gs98irx/croots/recipe/ijson_1655384050388/work
jmespath==0.10.0
openai==1.6.0
oss2==2.17.0
Pillow==9.5.0
pipreqs==0.4.11
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.18.0
pydantic==2.5.2
pydantic_core==2.14.5
PyMySQL==1.0.2
python-dateutil==2.8.2
redis==3.5.3
requests==2.28.2
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
starlette==0.27.0
supervisor==4.2.5
tinycss2==1.2.1
tomli==2.0.1
tqdm==4.65.0
typing_extensions==4.9.0
urllib3==1.26.14
uvicorn @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_d80hhx7s7h/croot/uvicorn-split_1678090080474/work
vine==5.0.0
webencodings==0.5.1
xlwt==1.3.0
yarg==0.1.9
zope.event==4.6
zope.interface==6.0
